requirements.txt
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
openai-whisper==20240930
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
ffmpeg-python==0.2.0
numpy==2.0.2